+++ OsitoPolarPlatform.API/bc-technicians/Interfaces/REST/Transform/TechnicianResourceFromEntityAssembler.cs
@@ -5,7 +5,7 @@ namespace OsitoPolarPlatform.API.bc_technicians.Interfaces.REST.Transform;
 
 public static class TechnicianResourceFromEntityAssembler
 {
-    public static TechnicianResource ToResourceFromEntity(Technician entity, double averageRating = 0.0) 
+    public static TechnicianResource ToResourceFromEntity(Technician entity, decimal averageRating = 0.0)
     {
         return new TechnicianResource
         (
